src/requirements.txt
matplotlib>=3.7.2
pandas>=2.0.3
numpy>=1.25.0
mne>=1.5.0
sklearn
tensorflow-macos